+++ pyanaconda/modules/payload/base/handler_base.py
@@ -70,7 +70,7 @@ class PayloadHandlerBase(KickstartBaseModule, metaclass=ABCMeta):
         the initialization task will run with the old list.
 
         :param sources: set a new sources
-        :type sources: instance of pyanaconda.modules.payload.base.source_base.PayloadSourceBase
+        :type sources: instance of pyanaconda.modules.payload.sources.source_base.PayloadSourceBase
         :raise: IncompatibleSourceError when source is not a supported type
                 SourceSetupError when attached sources are initialized
         """
