+++ pyanaconda/modules/payload/sources/live_os/live_os.py
@@ -24,7 +24,7 @@ from pyanaconda.core.constants import INSTALL_TREE
 from pyanaconda.core.signal import Signal
 from pyanaconda.core.util import execWithCapture
 from pyanaconda.modules.payload.constants import SourceType
-from pyanaconda.modules.payload.base.source_base import PayloadSourceBase
+from pyanaconda.modules.payload.sources.source_base import PayloadSourceBase
 from pyanaconda.modules.payload.sources.live_os.live_os_interface import \
     LiveOSSourceInterface
 from pyanaconda.modules.payload.sources.live_os.initialization import \
