+++ pyanaconda/modules/payload/sources/live_os/live_os_interface.py
@@ -21,7 +21,7 @@ from pyanaconda.dbus.interface import dbus_interface
 from pyanaconda.dbus.typing import *  # pylint: disable=wildcard-import
 from pyanaconda.dbus.property import emits_properties_changed
 from pyanaconda.modules.common.constants.interfaces import PAYLOAD_SOURCE_LIVE_OS
-from pyanaconda.modules.payload.base.source_base_interface import PayloadSourceBaseInterface
+from pyanaconda.modules.payload.sources.source_base_interface import PayloadSourceBaseInterface
 
 
 @dbus_interface(PAYLOAD_SOURCE_LIVE_OS.interface_name)
