+++ pyanaconda/modules/payload/sources/source_base.py
