+++ pyanaconda/modules/payload/sources/source_base_interface.py
